requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.9.1
werkzeug==3.0.1
anthropic==0.18.1
python-dotenv==1.0.0
email-validator==2.1.0
psycopg2-binary==2.9.9
stripe==7.8.0
APScheduler==3.10.4
